+++ CHANGELOG.md
@@ -72,6 +72,7 @@ All adequations have been done for all tests and existing scripts.
 - Improve step translator.
 - Delete inplace methods: rotation, translation and frame_mapping
 - OpenShell3D: faces_graph.
+- RevolutionSurface3D: Improve init and methods
 
 
 ### Changed